+++ pdfviewer/src/androidMain/kotlin/tech/kotlinlang/pdfviewer/PdfViewer.android.kt
@@ -4,7 +4,8 @@ import android.graphics.Bitmap
 import android.graphics.Bitmap.createBitmap
 import android.graphics.pdf.PdfRenderer
 import android.os.ParcelFileDescriptor
-import android.widget.ImageView
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
@@ -13,8 +14,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.asImageBitmap
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.viewinterop.AndroidView
 import java.io.File
 
 @Composable
@@ -38,20 +40,17 @@ internal actual fun PdfViewerLoaded(
     ) {
         val currentPdfRenderer = pdfRenderer ?: return@LazyColumn
         items(currentPdfRenderer.pageCount) { pageIndex ->
-            AndroidView(
-                modifier = modifier,
-                factory = { context ->
-                    val page = currentPdfRenderer.openPage(pageIndex)
-
-                    val bitmap = createBitmap(page.width, page.height, Bitmap.Config.ARGB_8888)
-                    page.render(bitmap, null, null, PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY)
-
-                    val imageView = ImageView(context)
-                    imageView.setImageBitmap(bitmap)
-
-                    page.close()
-                    imageView
-                },
+            val bitmap = remember {
+                val page = currentPdfRenderer.openPage(pageIndex)
+                val bitmap = createBitmap(page.width, page.height, Bitmap.Config.ARGB_8888)
+                page.render(bitmap, null, null, PdfRenderer.Page.RENDER_MODE_FOR_DISPLAY)
+                bitmap
+            }
+            Image(
+                modifier = Modifier.fillMaxWidth(),
+                bitmap = bitmap.asImageBitmap(),
+                contentScale = ContentScale.FillWidth,
+                contentDescription = null,
             )
         }
     }